Setup Program to load a file, create the drawing and flush it to JPG. Since this API is experimental, temporarily throwing TDD overboard to get something visually on the screen.

+++ src/FoldingRenderer/Domain/Types/PanelCrease.cs
@@ -16,9 +16,9 @@ namespace FoldingRenderer.Domain.Types {
       Right = right;
     }
 
-    public PanelCrease WithBottom(int bottom) => new PanelCrease(bottom, Top, Left, Right);
-    public PanelCrease WithTop(int top) => new PanelCrease(Bottom, top, Left, Right);
-    public PanelCrease WithLeft(int left) => new PanelCrease(Bottom, Top, left, Right);
-    public PanelCrease WithRight(int right) => new PanelCrease(Bottom, Top, Left, right);
+    public PanelCrease WithBottom(double bottom) => new PanelCrease(bottom, Top, Left, Right);
+    public PanelCrease WithTop(double top) => new PanelCrease(Bottom, top, Left, Right);
+    public PanelCrease WithLeft(double left) => new PanelCrease(Bottom, Top, left, Right);
+    public PanelCrease WithRight(double right) => new PanelCrease(Bottom, Top, Left, right);
   }
 }
